new ada benchmark using CountMinSketch on IP adresses instead of DD_Sketch

+++ EDADS/src/main/java/Benchmark/Sources/SimpleQuerySource.java
@@ -6,7 +6,7 @@ import org.apache.flink.streaming.api.windowing.time.Time;
 
 import java.util.Random;
 
-public class SimpleQuerySource extends RichParallelSourceFunction<Double> {
+public class SimpleQuerySource extends RichParallelSourceFunction<Integer> {
 
     private final long runtime;
     private final int throughput;
@@ -19,7 +19,7 @@ public class SimpleQuerySource extends RichParallelSourceFunction<Double> {
     }
 
     @Override
-    public void run(SourceContext<Double> ctx) throws Exception {
+    public void run(SourceContext<Integer> ctx) throws Exception {
         Random random = new Random();
 
         long startTs = System.currentTimeMillis();
@@ -34,7 +34,7 @@ public class SimpleQuerySource extends RichParallelSourceFunction<Double> {
             long time = System.currentTimeMillis();
 
             for (int i = 0; i < throughput; i++) {
-                ctx.collectWithTimestamp(random.nextDouble(), time);
+                ctx.collectWithTimestamp(random.nextInt(), time);
             }
 
             while (System.currentTimeMillis() < time + 1000) {
