+++ EDADS/src/main/java/Benchmark/Sources/StratifiedTimestampedQuerySource.java
@@ -8,7 +8,7 @@ import org.apache.flink.streaming.api.windowing.time.Time;
 
 import java.util.Random;
 
-public class StratifiedTimestampedQuerySource extends RichParallelSourceFunction<Tuple2<Integer, TimestampedQuery<Double>>> {
+public class StratifiedTimestampedQuerySource extends RichParallelSourceFunction<Tuple2<Integer, TimestampedQuery<Integer>>> {
 
     private final int throughput;
     private final long wait;
@@ -26,7 +26,7 @@ public class StratifiedTimestampedQuerySource extends RichParallelSourceFunction
     }
 
     @Override
-    public void run(SourceContext<Tuple2<Integer, TimestampedQuery<Double>>> ctx) throws Exception {
+    public void run(SourceContext<Tuple2<Integer, TimestampedQuery<Integer>>> ctx) throws Exception {
         Random random = new Random();
 
         long startTs = System.currentTimeMillis();
@@ -44,7 +44,7 @@ public class StratifiedTimestampedQuerySource extends RichParallelSourceFunction
             for (int i = 0; i < throughput; i++) {
                 int relativeTimestamp = random.nextInt(synopsisRuntime);
                 long query_timestamp = startTs + relativeTimestamp * 1000;
-                TimestampedQuery<Double> timestampedQuery = new TimestampedQuery<Double>(random.nextDouble(), query_timestamp);
+                TimestampedQuery<Integer> timestampedQuery = new TimestampedQuery<Integer>(random.nextInt(), query_timestamp);
                 ctx.collectWithTimestamp(new Tuple2<>(random.nextInt(stratification), timestampedQuery), time);
             }
